Update unit test
